Rename tests to in_diff
