+++ tests/cli/in_diff.rs
